user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
alembic
python-dotenv
pytest
httpx
python-jose[cryptography]